+++ hata/discord/webhook/webhook_base.py
@@ -22,6 +22,8 @@ class WebhookBase(UserBase):
         The webhook's avatar's hash in `uint128`.
     avatar_type : ``IconType``
         The webhook's avatar's type.
+    banner_color : `None` or ``Color``
+        The user's banner color if has any.
     banner_hash : `int`
         The user's banner's hash in `uint128`.
     banner_type : ``IconType``